removed some errors
Fixes to vec3, removed reliance on old mojang vector library, fixed some issues with the peripheral provider, but it still doesn't work. Fixed DataHelpers and A couple other things. For future reference items and blocks need to be added back to the creative tab. ItemProperties where changed so assigning a creative tab is different.